Move You.com tool into extensions package
